0.6.2

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ChangePasswordRequest;
 use App\Http\Requests\CreateUserRequest;
+use App\Http\Requests\DeleteUserRequest;
 use App\Http\Requests\LoginRequest;
 use App\Http\Resources\UserCollection;
 use App\Http\Resources\UserResource;
@@ -18,6 +19,8 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Symfony\Component\HttpFoundation\Exception\BadRequestException;
+use Symfony\Component\HttpKernel\Exception\LockedHttpException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 
 class UserController extends Controller
@@ -28,7 +31,7 @@ class UserController extends Controller
 
     private function generate_code($length)
     {
-        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ._+#%@-';
+        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
         $code = '';
 
         for ($i = 0; $i < $length; $i++) {
@@ -52,13 +55,11 @@ class UserController extends Controller
 
             $new_user = User::create([
                 'company_id' => $company_id,
-                'code' => $request->code,
                 'user_name' => $user_name,
                 'name' => $request->name,
                 'state' => 'I',
                 'phone_number' => $request->phoneNumber ?? null,
                 'created_at' => Carbon::now(),
-
             ]);
 
             foreach ($request->roles as $role) {
@@ -79,7 +80,7 @@ class UserController extends Controller
                 'company_id' => $company_id,
                 'user_id' => $sender->id,
                 'token_id' => $sender->currentAccessToken()->id,
-                'action' => 'Created user ' . $user_name,
+                'action' => 'Created user '.$user_name,
                 'occured_at' => Carbon::now(),
             ]);
 
@@ -110,16 +111,21 @@ class UserController extends Controller
                 'user_name' => $request->userName,
             ]);
 
-            if (!$user) {
+            if (! $user) {
                 throw new UnauthorizedHttpException(random_bytes(32));
             }
 
+            $currentAttempts = $user->attempts;
             $user->last_active = Carbon::now();
             $user->save();
 
+            if ($currentAttempts > 2) {
+                throw new LockedHttpException();
+            }
+
             $password = $user->passwords()->orderBy('set_time', 'desc')->first();
 
-            if (!Hash::check($request->password, $password->password)) {
+            if (! Hash::check($request->password, $password->password)) {
                 Log::insert([
                     'company_id' => $user->company_id,
                     'user_id' => $user->id,
@@ -127,6 +133,9 @@ class UserController extends Controller
                     'action' => 'Tried to log in with a wrong password',
                     'occured_at' => Carbon::now(),
                 ]);
+
+                $user->attempts = $currentAttempts + 1;
+                $user->save();
 
                 throw new UnauthorizedHttpException(random_bytes(32));
             }
@@ -171,6 +180,7 @@ class UserController extends Controller
             if (
                 $e instanceof UnauthorizedHttpException
                 || $e instanceof AuthorizationException
+                || $e instanceof LockedHttpException
             ) {
                 throw $e;
             }
@@ -214,10 +224,33 @@ class UserController extends Controller
             $sender->last_active = Carbon::now();
             $sender->save();
 
-            $oldPasswordIds = $sender->passwords()->limit(100)->offset(10)->pluck('id');
+            $senderRoles = array_map(
+                fn ($role) => $role['role'],
+                $sender->roles->toArray()
+            );
+
+            $isSenderAdmin = in_array('AM', $senderRoles);
+
+            if ($request->userName && ! $isSenderAdmin) {
+                throw new UnauthorizedHttpException(random_bytes(32));
+            }
+
+            if ($isSenderAdmin && $request->userName) {
+                $subjectOfChange = User::firstWhere([
+                    'user_name' => $request->userName,
+                ]);
+
+                if (! $subjectOfChange) {
+                    throw new NotFoundHttpException();
+                }
+            } else {
+                $subjectOfChange = $sender;
+            }
+
+            $oldPasswordIds = $subjectOfChange->passwords()->limit(100)->offset(10)->pluck('id');
             UserPassword::whereIn('id', $oldPasswordIds)->delete();
 
-            $previousPasswords = $sender->passwords;
+            $previousPasswords = $subjectOfChange->passwords;
             $newPassword = $request->password;
 
             foreach ($previousPasswords as $previousPassword) {
@@ -230,24 +263,25 @@ class UserController extends Controller
                 }
             }
 
-            $sender->tokens()->delete();
-            $sender->passwords()->create([
+            $subjectOfChange->attempts = 0;
+            $subjectOfChange->tokens()->delete();
+            $subjectOfChange->passwords()->create([
                 'password' => Hash::make($newPassword),
                 'is_generated' => 0,
                 'set_time' => Carbon::now(),
             ]);
-            $sender->save();
+            $subjectOfChange->save();
 
-            $roles = array_map(
+            $subjectRoles = array_map(
                 fn ($role) => $role['role'],
-                $sender->roles->toArray()
+                $subjectOfChange->roles->toArray()
             );
-            $token = $sender->createToken('boreal', $roles);
+            $token = $subjectOfChange->createToken('boreal', $subjectRoles);
             $tokenExpiration = Carbon::now()->addHours(25)->toDateTimeString();
 
             Log::insert([
-                'company_id' => $sender->company_id,
-                'user_id' => $sender->id,
+                'company_id' => $subjectOfChange->company_id,
+                'user_id' => $subjectOfChange->id,
                 'token_id' => 0,
                 'action' => 'Changed password',
                 'occured_at' => Carbon::now(),
@@ -255,7 +289,7 @@ class UserController extends Controller
 
             $userResource = new UserResource($sender->load('company'));
 
-            return array_merge(
+            return $sender->id === $subjectOfChange->id ? array_merge(
                 $userResource->toArray(0),
                 [
                     'token' => [
@@ -265,8 +299,15 @@ class UserController extends Controller
                         'expiresAt' => $tokenExpiration,
                     ],
                 ]
-            );
+            ) : $userResource;
         } catch (Exception $e) {
+            if (
+                $e instanceof NotFoundHttpException
+                || $e instanceof UnauthorizedHttpException
+            ) {
+                throw $e;
+            }
+
             throw new BadRequestException();
         }
     }
@@ -284,7 +325,7 @@ class UserController extends Controller
                 'company_id' => $sender->company_id,
                 'user_id' => $sender->id,
                 'token_id' => $sender->currentAccessToken()->id,
-                'action' => 'Accessed ' . $companyUsers->count() . ' users',
+                'action' => 'Accessed '.$companyUsers->count().' users',
                 'occured_at' => Carbon::now(),
             ]);
 
@@ -294,15 +335,21 @@ class UserController extends Controller
         }
     }
 
-    public function remove(int $id)
+    public function delete(DeleteUserRequest $request)
     {
         try {
             $sender = request()->user();
             $sender->last_active = Carbon::now();
             $sender->save();
 
-            $user = $sender->company->users()->findOrFail($id);
+            $user = User::firstWhere([
+                'user_name' => $request->userName,
+            ]);
             $this->authorize('remove', $user);
+
+            if ($sender->id === $user->id) {
+                throw new UnauthorizedHttpException(random_bytes(32));
+            }
 
             $user->delete();
 
@@ -310,7 +357,7 @@ class UserController extends Controller
                 'company_id' => $sender->company_id,
                 'user_id' => $sender->id,
                 'token_id' => $sender->currentAccessToken()->id,
-                'action' => 'Removed user ' . $user->id,
+                'action' => 'Removed user '.$user->id,
                 'occured_at' => Carbon::now(),
             ]);
         } catch (Exception $e) {
